requirements.txt
docutils==0.14
elasticsearch==6.3.1
lockfile==0.12.2
pika==0.13.0
python-dateutil==2.8.0
python-dotenv==0.10.1
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
